Introduced EventProcessingConfiguration as ModuleConfiguration.
