+++ core/src/main/java/org/axonframework/config/DefaultConfigurer.java
@@ -83,8 +83,6 @@ public class DefaultConfigurer implements Configurer {
 
     private final Configuration config = new ConfigurationImpl();
 
-    private final Component<EventProcessorRegistry> eventProcessorRegistry;
-
     private final MessageMonitorFactoryBuilder messageMonitorFactoryBuilder = new MessageMonitorFactoryBuilder();
     private final Component<BiFunction<Class<?>, String, MessageMonitor<Message<?>>>> messageMonitorFactoryComponent =
             new Component<>(config, "monitorFactory", messageMonitorFactoryBuilder::build);
@@ -126,11 +124,6 @@ public class DefaultConfigurer implements Configurer {
         components.put(QueryGateway.class, new Component<>(config, "queryGateway", this::defaultQueryGateway));
         components.put(ResourceInjector.class,
                        new Component<>(config, "resourceInjector", this::defaultResourceInjector));
-
-        eventProcessorRegistry = new Component<>(config,
-                                                 "eventProcessorRegistry",
-                                                 c -> defaultEventProcessorRegistry());
-        components.put(EventProcessorRegistry.class, eventProcessorRegistry);
     }
 
     /**
@@ -261,15 +254,6 @@ public class DefaultConfigurer implements Configurer {
     }
 
     /**
-     * Provides the default implementation of {@link EventProcessorRegistry}.
-     *
-     * @return the default implementation of {@link EventProcessorRegistry}
-     */
-    protected EventProcessorRegistry defaultEventProcessorRegistry() {
-        return new DefaultEventProcessorRegistry();
-    }
-
-    /**
      * Provides the default EventBus implementation. Subclasses may override this method to provide their own default.
      *
      * @param config The configuration based on which the component is initialized.
@@ -320,13 +304,6 @@ public class DefaultConfigurer implements Configurer {
     public Configurer configureCorrelationDataProviders(
             Function<Configuration, List<CorrelationDataProvider>> correlationDataProviderBuilder) {
         correlationProviders.update(correlationDataProviderBuilder);
-        return this;
-    }
-
-    @Override
-    public Configurer configureEventProcessorRegistry(
-            Function<Configuration, EventProcessorRegistry> eventProcessorRegistryBuilder) {
-        eventProcessorRegistry.update(eventProcessorRegistryBuilder);
         return this;
     }
 
@@ -410,6 +387,11 @@ public class DefaultConfigurer implements Configurer {
     @Override
     public Configuration buildConfiguration() {
         if (!initialized) {
+            boolean missingEventProcessingConfiguration = modules.stream()
+                                                                 .noneMatch(m -> m.instance() instanceof EventProcessingConfiguration);
+            if (missingEventProcessingConfiguration) {
+                registerModule(new EventProcessingConfiguration());
+            }
             invokeInitHandlers();
         }
         return config;
@@ -422,7 +404,6 @@ public class DefaultConfigurer implements Configurer {
     protected void invokeInitHandlers() {
         initialized = true;
         initHandlers.stream().sorted(comparingInt(ConsumerHandler::phase)).forEach(h -> h.accept(config));
-        eventProcessorRegistry.get().initialize(config);
     }
 
     /**
@@ -430,7 +411,6 @@ public class DefaultConfigurer implements Configurer {
      */
     protected void invokeStartHandlers() {
         startHandlers.stream().sorted(comparingInt(RunnableHandler::phase)).forEach(RunnableHandler::run);
-        eventProcessorRegistry.get().start();
     }
 
     /**
@@ -438,7 +418,6 @@ public class DefaultConfigurer implements Configurer {
      */
     protected void invokeShutdownHandlers() {
         shutdownHandlers.stream().sorted(comparingInt(RunnableHandler::phase).reversed()).forEach(RunnableHandler::run);
-        eventProcessorRegistry.get().shutdown();
     }
 
     /**
