+++ core/src/test/java/org/axonframework/config/DefaultConfigurerTest.java
@@ -51,6 +51,7 @@ import static org.axonframework.commandhandling.model.AggregateLifecycle.apply;
 import static org.axonframework.common.AssertUtils.assertWithin;
 import static org.axonframework.config.AggregateConfigurer.defaultConfiguration;
 import static org.axonframework.config.AggregateConfigurer.jpaMappedConfiguration;
+import static org.axonframework.config.ConfigAssertions.assertExpectedModules;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
@@ -85,7 +86,10 @@ public class DefaultConfigurerTest {
         config.commandBus().dispatch(GenericCommandMessage.asCommandMessage("test"), callback);
         assertEquals("test", callback.get());
         assertNotNull(config.repository(StubAggregate.class));
-        assertEquals(1, config.getModules().size());
+        assertEquals(2, config.getModules().size());
+        assertExpectedModules(config,
+                              AggregateConfiguration.class,
+                              EventProcessingConfiguration.class);
     }
 
     @Test
@@ -95,7 +99,7 @@ public class DefaultConfigurerTest {
                                                                    c -> TrackingEventProcessorConfiguration
                                                                            .forParallelProcessing(2))
                                                 .configureEmbeddedEventStore(c -> new InMemoryEventStorageEngine())
-                                                .configureEventProcessorRegistry(c -> new DefaultEventProcessorRegistry()
+                                                .registerModule(new EventProcessingConfiguration()
                                                         .usingTrackingProcessors())
                                                 .registerModule(
                                                         new EventHandlingConfiguration()
@@ -104,8 +108,8 @@ public class DefaultConfigurerTest {
                                                 )
                                                 .start();
         try {
-            TrackingEventProcessor processor = config.eventProcessorRegistry().eventProcessor(getClass().getPackage().getName(), TrackingEventProcessor.class)
-                                                                                                        .orElseThrow(RuntimeException::new);
+            TrackingEventProcessor processor = config.eventProcessingConfiguration().eventProcessor(getClass().getPackage().getName(), TrackingEventProcessor.class)
+                                                     .orElseThrow(RuntimeException::new);
             assertWithin(5, TimeUnit.SECONDS, () -> assertEquals(2, config.getComponent(TokenStore.class).fetchSegments(processor.getName()).length));
         } finally {
             config.shutdown();
@@ -126,8 +130,8 @@ public class DefaultConfigurerTest {
                                                 )
                                                 .start();
         try {
-            TrackingEventProcessor processor = config.eventProcessorRegistry().eventProcessor(getClass().getPackage().getName(), TrackingEventProcessor.class)
-                                                                                                        .orElseThrow(RuntimeException::new);
+            TrackingEventProcessor processor = config.eventProcessingConfiguration().eventProcessor(getClass().getPackage().getName(), TrackingEventProcessor.class)
+                                                     .orElseThrow(RuntimeException::new);
             assertWithin(5, TimeUnit.SECONDS, () -> assertEquals(2, config.getComponent(TokenStore.class).fetchSegments(processor.getName()).length));
         } finally {
             config.shutdown();
@@ -176,7 +180,10 @@ public class DefaultConfigurerTest {
         config.commandBus().dispatch(GenericCommandMessage.asCommandMessage("test"), callback);
         assertEquals("test", callback.get());
         assertNotNull(config.repository(StubAggregate.class));
-        assertEquals(1, config.getModules().size());
+        assertEquals(2, config.getModules().size());
+        assertExpectedModules(config,
+                              EventProcessingConfiguration.class,
+                              AggregateConfiguration.class);
         verify(transactionManager).startTransaction();
     }
 
@@ -197,7 +204,9 @@ public class DefaultConfigurerTest {
         config.commandBus().dispatch(GenericCommandMessage.asCommandMessage("test"), callback);
         assertEquals("test", callback.get());
         assertNotNull(config.repository(StubAggregate.class));
-        assertEquals(1, config.getModules().size());
+        assertTrue(config.getModules()
+                         .stream()
+                         .anyMatch(m -> m instanceof AggregateConfiguration));
         verify(transactionManager).startTransaction();
     }
 
@@ -242,7 +251,10 @@ public class DefaultConfigurerTest {
         config.commandBus().dispatch(GenericCommandMessage.asCommandMessage("test"), callback);
         assertEquals("test", callback.get());
         assertNotNull(config.repository(StubAggregate.class));
-        assertEquals(1, config.getModules().size());
+        assertEquals(2, config.getModules().size());
+        assertExpectedModules(config,
+                              EventProcessingConfiguration.class,
+                              AggregateConfiguration.class);
         verify(transactionManager).startTransaction();
     }
 
@@ -275,7 +287,12 @@ public class DefaultConfigurerTest {
                                                 .configureEmbeddedEventStore(c -> new InMemoryEventStorageEngine())
                                                 .start();
 
-        assertThat(config.getModules().size(), CoreMatchers.is(3));
+        assertThat(config.getModules().size(), CoreMatchers.is(4));
+        assertExpectedModules(config,
+                              EventHandlingConfiguration.class,
+                              EventHandlingConfiguration.class,
+                              AggregateConfiguration.class,
+                              EventProcessingConfiguration.class);
     }
 
     @Test
